Layout: remove destination object from decode method
This relates to issue #8 in that it provided a way to hide the true
structure by allowing fields within unions (but not structs) to be
placed in the parent object.  This is too confusing: the rule is that
Layout.decode always returns the decoded value; if that value is a
Union, then even if the variant in it is a number what gets returned is
an Object not the raw number.