passim: replace instanceof with Buffer.isBuffer
instanceOf check may fail in browser environment where multiple
versions of the buffer package may be bundled together, hence
resulting in different instances of Buffer being present in the
code. A more resilient way of checking whether the parameter supplied
is a buffer is the Buffer.isBuffer() method which I'm migrating to in
this PR.

From #24.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>