Layout: remove dependency on assert module
Apparently some environments where this package is being used do not
support an assert API.  The only use of the API was a sanity check
that unnamed fields were defined with a fixed length.  A comment
noted, and inspection confirms, that the Structure constructor
explicitly rejects attempts to define such a field.  So just drop the
check.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>